Loading/saving customers and accounts

+++ BankApp/BankObjects/Identified.cs
@@ -1,13 +1,15 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.Serialization;
+using BankApp.IO;
 
 namespace BankApp.BankObjects
 {
 	public abstract class Identified
 	{
-		public long ID { get; protected set; }
+		public uint ID { get; protected set; }
 
-		public static long GenerateUniqueID(IEnumerable<Identified> existing)
+		public static uint GenerateUniqueID(IEnumerable<Identified> existing)
 		{
 			return existing.Select(i => i.ID).Max() + 1;
 		}
